feat(editor): add reusable markdown editor component
